chore(vscode): rename download-copy-sqlite-esbuild.js to download-copy-sqlite.js and update imports
